removed old classical ttt
